Fix missing dtype cast

+++ src/peft/tuners/lora/dora.py
@@ -94,7 +94,7 @@ class DoraLinearLayer(nn.Module):
     def get_lora_weight(self, lora_A, lora_B, adapter_name: Optional[str] = None):
         # Don't use `lora_weight = lora_B.weight @ lora_A.weight` because this causes errors with FSDP. Instead,
         # calculate the same but using forward.
-        x_eye = torch.eye(lora_A.weight.shape[1], device=lora_A.weight.device)
+        x_eye = torch.eye(lora_A.weight.shape[1], device=lora_A.weight.device, dtype=lora_A.weight.dtype)
         lora_weight = lora_B(lora_A(x_eye)).T
         return lora_weight
